==> 10_alignment.R <==
library(here)
library(tidyverse)
library(energy)
library(jsonlite)

args <- commandArgs(trailingOnly = TRUE)

dists.file <- args[1]
labels.file <- args[2]
output.dir <- args[3]

pairwise.dists <-
  as.matrix(read_table(here(dists.file), col_names = FALSE))

  # Read the JSON content as a character string, replace NaNs with "NaN" (as a string), and parse it
json_content <- readLines(here(labels.file), warn = FALSE)
json_text <- paste(json_content, collapse = "\n")
modified_json_text <- gsub("\\bNaN\\b", '"NaN"', json_text)
parsed_json <- fromJSON(modified_json_text, simplifyVector = TRUE)
signal.labels <- as.data.frame(parsed_json) %>%
  rename(
    game = V1,
    speaker = V2,
    referent = V3,
    referent_id = V4
  )

# signal.labels <-
#   as.data.frame(read_json(here("outputs/all_signal_labels.json"), simplifyVector = TRUE)) %>%
#   rename(
#     game = V1,
#     speaker = V2,
#     referent = V3,
#     referent_id = V4
#   )
wcs <- read_json(here("stim/wcs_row_F.json"))

speaker.ids <- unique(signal.labels$speaker)
game.ids <- unique(signal.labels$game)

alignments <- matrix(ncol = 2, nrow = 0)


# Calculate how similar the signal distances are between the two speakers in each game
# Loop through each game. For each speaker in the game, make one matrix of signal distances
# Then, calculate distance correlation between the matrices
for (id in game.ids) {
  if (id == "NaN") {
    next
  }

  # Find the speaker ids for this game
  game.speaker.ids <-
    unique(signal.labels$speaker[signal.labels$game == id])

  speaker1.signal.indices <-
    which(signal.labels$speaker %in% game.speaker.ids[1])

  speaker2.signal.indices <-
    which(signal.labels$speaker %in% game.speaker.ids[2])

  # make array of available signal indices for each speaker
  # signals.speaker1 <-
  #   as.integer(unique(signal.labels$referent_id[signal.labels$speaker == game.speaker.ids[1]]))
  # signals.speaker2 <-
  #   as.integer(unique(signal.labels$referent_id[signal.labels$speaker == game.speaker.ids[2]]))

  n.signals <- 40


  # Make an empty vector of signal distances for each speaker
  game.signal.dists <- numeric(0)

  # For each signal for speaker 1, find the corresponding signal for speaker 2
  # Then find distance between those two signals
  for (i in 1:n.signals) {
    # find the referent id for the signal for speaker 1
    speaker1.signal.referent_id <-
      signal.labels$referent_id[speaker1.signal.indices[i]]
    # find the signal for speaker 2 that corresponds to the same referent
    speaker2.signal.index <-
      speaker2.signal.indices[signal.labels$referent_id[speaker2.signal.indices] == speaker1.signal.referent_id]


    # find the distance between the two signals
    speakers.signal.dist <-
      pairwise.dists[speaker1.signal.indices[i], speaker2.signal.index]


    # find the distance between the two signals
    # speakers.signal.dist <-
    #   pairwise.dists[speaker1.signal.indices[i], speaker2.signal.indices[i]]

    # add the distance to the vector of distances
    game.signal.dists <-
      c(game.signal.dists, speakers.signal.dist)


  }
  # Calculate the mean of the list of distances, make sure to exclude NaNs

  alignment <-
    mean(game.signal.dists, na.rm = TRUE)
  alignments <- rbind(alignments, c(id, alignment))
}

colnames(alignments) <- c("game", "dist")

# Normalize s.t. 0 is least aligned and 1 is most aligned



write.csv(alignments, here(paste0(output.dir, "/alignments.csv")), row.names = FALSE)
